+++ packages/api/app.js
@@ -15,7 +15,7 @@ const
   restify = require('restify'),
   { fullResponse, queryParser, jsonBodyParser, pre: { sanitizePath } } = restify.plugins,
   config = require('config'),
-  { preRequest, preCors, Honeybadger, getVersion, postToSlack } = require('./lib/helpers/apiUtils'),
+  { preRequest, preCors, Honeybadger, getVersion, postToMattermost } = require('./lib/helpers/apiUtils'),
   routes = require('./lib/routes'),
   bunyan = require('bunyan');
 
@@ -23,7 +23,9 @@ const log = bunyan.createLogger({ name: 'opensensemap-api', serializers: bunyan.
 
 const server = restify.createServer({
   name: `opensensemap-api (${getVersion})`,
-  log
+  log,
+  onceNext: true,
+  strictNext: false,
 });
 
 // We're using caddy as proxy. It supplies a 'X-Forwarded-Proto' header
@@ -48,7 +50,7 @@ db.connect()
     // start the server
     server.listen(Number(config.get('port')), function () {
       log.info(`${server.name} listening at ${server.url}`);
-      postToSlack(`openSenseMap API started. Version: ${getVersion}`);
+      postToMattermost(`openSenseMap API started. Version: ${getVersion}`);
     });
   })
   .catch(function (err) {